% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afscISS-package.R
\docType{package}
\name{afscISS-package}
\alias{afscISS}
\alias{afscISS-package}
\title{afscISS: Retrieve composition data ISS}
\description{
Retrieve bootstrap estimates of age/length/conditional age-at-length input sample size produced by the surveyISS package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/afsc-assessments/afscISS/}
  \item \url{https://benwilliams-noaa.github.io/surveyISS/}
  \item Report bugs at \url{https://github.com/afsc-assessments/afscISS/issues}
}

}
\author{
\strong{Maintainer}: Pete Hulson \email{pete.hulson@noaa.gov} (\href{https://orcid.org/000-0001-5651-2740}{ORCID})

Authors:
\itemize{
  \item Ben Williams \email{ben.williams@noaa.gov} (\href{https://orcid.org/000-0001-7295-2076}{ORCID})
}

}
\keyword{internal}
