% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_iss}
\alias{data_iss}
\title{Results from surveyISS package}
\format{
A list
\describe{
A list of ISS, age/length pop'n numbers, bootstrap bias, mean length-at-age, replicated rss, replicated age/length pop'n numbers
}
}
\usage{
data_iss
}
\description{
A list containing the various output dataframes from the surveyISS package
}
\keyword{datasets}
