% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_bias}
\alias{get_bias}
\title{Get stock-specific bootstrap bias results}
\usage{
get_bias(
  species = 21720,
  region = "goa",
  comp = "age",
  sex_cat = 4,
  spec_case = NULL
)
}
\arguments{
\item{species}{AFSC species code (default = 21720, pacific cod)}

\item{region}{survey region. options are 'ai', 'ebs', 'ebs_slope', 'goa', and 'nebs' (default = 'goa')}

\item{comp}{type of composition for which ISS desired. options are 'age', 'length', and 'caal'}

\item{sex_cat}{sex category for which bootstrap bias desired, options are 0 (sexes combined pre-expansion), 1 (males), 2 (females), 12 (males-female comp that sums to 1), and 4 (sexes combined post-expansion) (default = 4)}

\item{spec_case}{description string if getting bootstrap bias for special case. options are 'ai_subreg', 'bsre', 'dr', 'rebs', 'w_c_egoa', 'w140', 'wc_egoa' (default = NULL)}
}
\value{
a dataframe of bootstrap bias statistics
}
\description{
Function that retrieves bootstrap bias results
}
