% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_comp_stats}
\alias{get_comp_stats}
\title{function to summarize bootstrap results for composition data}
\usage{
get_comp_stats(data, grp, column, dgts)
}
\arguments{
\item{data}{dataframe to be summarized}

\item{grp}{grouping for summarization}

\item{column}{column name to be summarized across iterations}

\item{dgts}{number of significant digits in rounding results}
}
\value{
dataframe with summary bootstrap statistics
}
\description{
Function that computes mean and percentiles for iterated composition results from the surveyISS package (NOTE: this is not a user fcn, this is a developer/maintainer fcn)
}
