% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_comps}
\alias{get_comps}
\title{Developer function to get composition data}
\usage{
get_comps(base_data, res_data, iter, tot, reg)
}
\arguments{
\item{base_data}{base dataframe without resampling}

\item{res_data}{resampled dataframe}

\item{iter}{which dataframe is being summarized (for print message)}

\item{tot}{total number of dataframes being summarized (for print message)}

\item{reg}{survey region in which dataframe is being summarized (for print message)}
}
\value{
dataframe with summary bootstrap statistics
}
\description{
Function that computes age/length/conditional age-at-length composition with bootstrap 95\% simulation intervals (NOTE: this is not a user fcn, this is a developer/maintainer fcn)
}
\keyword{internal}
