% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_popn}
\alias{get_popn}
\title{Get stock-specific composition results}
\usage{
get_popn(
  species = 21720,
  region = "goa",
  comp = "age",
  sex_cat = 4,
  spec_case = NULL
)
}
\arguments{
\item{species}{AFSC species code (default = 21720, pacific cod)}

\item{region}{survey region. options are 'ai', 'ebs', 'ebs_slope', 'goa', and 'nebs' (default = 'goa')}

\item{comp}{type of pop'n numbers desired, options are 'age' and 'length'}

\item{sex_cat}{sex category for which pop'n numbers desired. options are 0 (sexes combined pre-expansion), 1 (males), 2 (females), 3 (unsexed), and 4 (sexes combined post-expansion) (default = 4)}

\item{spec_case}{description string if getting pop'n numbers for special case. options are 'ai_subreg', 'bsre', 'dr', 'rebs', 'w_c_egoa', 'w140', 'wc_egoa' (default = NULL)}
}
\value{
a dataframe of age or length pop'n numbers (note that age data also include mean and sd in length-at-age)
}
\description{
Function that retrieves age and length pop'n numbers for AFSC stock assessments
}
