% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_RSS}
\alias{get_RSS}
\title{Get stock-specific RSS results}
\usage{
get_RSS(
  species = 21720,
  region = "goa",
  comp = "age",
  sex_cat = 4,
  spec_case = NULL
)
}
\arguments{
\item{species}{AFSC species code (default = 21720, pacific cod)}

\item{region}{survey region. options are 'ai', 'ebs', 'ebs_slope', 'goa', and 'nebs' (default = 'goa')}

\item{comp}{type of composition for which ISS desired. options are 'age', 'length', and 'caal'}

\item{sex_cat}{sex category for which composition ISS desired. options are 0, 1, 2, 12, and 4 (default = 4)}

\item{spec_case}{description string if getting ISS for special case. options are 'ai_subreg', 'bsre', 'dr', 'rebs', 'w_c_egoa', 'w140', 'wc_egoa' (default = NULL)}
}
\value{
a dataframe of composition ISS
}
\description{
Function that retrieves composition replicates of Realized Sample Size results for AFSC stock assessments
}
