% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_stats}
\alias{get_stats}
\title{function to summarize bootstrap results}
\usage{
get_stats(data, grp, column)
}
\arguments{
\item{data}{dataframe to be summarized}

\item{grp}{grouping for summarization}

\item{column}{column name to be summarized across iterations}
}
\value{
dataframe with summary bootstrap statistics
}
\description{
Function that computes mean and percentiles for iterated (comps and rss) results from the surveyISS package (NOTE: this is not a user fcn, this is a developer/maintainer fcn)
}
