% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_surveyISS}
\alias{get_surveyISS}
\title{Developer function to pull surveyISS results}
\usage{
get_surveyISS(region)
}
\arguments{
\item{region}{region for surveyISS output to be gathered}
}
\value{
dataframe with summary results from surveyISS
}
\description{
Function that retrieves results from the surveyISS package (NOTE: this is not a user fcn, this is a developer/maintainer fcn)
}
\keyword{internal}
