% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{iss_stats}
\alias{iss_stats}
\title{Developer function to get input sample size stats}
\usage{
iss_stats(iss_data, rss_data, iter, tot, reg)
}
\arguments{
\item{iss_data}{dataframe of iss values}

\item{rss_data}{dataframe of rss values}

\item{iter}{which dataframe is being summarized (for print message)}

\item{tot}{total number of dataframes being summarized (for print message)}

\item{reg}{survey region in which dataframe is being summarized (for print message)}
}
\value{
dataframe with summary bootstrap statistics
}
\description{
Function that computes the SD for input sample size (NOTE: this is not a user fcn, this is a developer/maintainer fcn)
}
\keyword{internal}
