% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pkg_data}
\alias{pkg_data}
\title{function to get surveyISS results transferred to data package}
\usage{
pkg_data(append = TRUE)
}
\arguments{
\item{append}{boolean. whether to append results to previous results (default = TRUE)}
}
\value{
.rda files within /data folder
}
\description{
Function that creates package data for afscISS (NOTE: this is not a user fcn, this is a developer/maintainer fcn)
}
