% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pkg_dta}
\alias{pkg_dta}
\title{function to get surveyISS results transferred to data package}
\usage{
pkg_dta(append = TRUE)
}
\arguments{
\item{append}{boolean. whether to append results to previous results (default = TRUE)}
}
\value{
.rda files within /data folder
}
\description{
Function that retrieves composition ISS results for AFSC stock assessments
}
