% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{summ_stats}
\alias{summ_stats}
\title{wrapper function to perform summary statistics}
\usage{
summ_stats(data, iter, tot, reg)
}
\arguments{
\item{data}{dataframe to be summarized}

\item{iter}{which dataframe is being summarized (for print message)}

\item{tot}{total number of dataframes being summarized (for print message)}

\item{reg}{survey region in which dataframe is being summarized (for print message)}
}
\value{
dataframe with summary bootstrap statistics
}
\description{
Function that tests whether summary statistics should be performed on results, and computes statistics with get_stats function
}
