% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tickr}
\alias{tickr}
\title{Adjust axis tick marks and labels}
\usage{
tickr(data, var, to = 5, start = NULL, end = NULL, min = NULL)
}
\arguments{
\item{data}{= input dataframe}

\item{var}{= variable of interest e.g., year}

\item{to}{= step increase desired e.g., every 5 years}

\item{start}{= adjust the start value}

\item{end}{= adjust the end vlaue}

\item{min}{= lowest value to label}
}
\description{
Adjust axis tick marks and labels
}
\keyword{internal}
