% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output.R
\name{get_RSS}
\alias{get_RSS}
\title{Get stock-specific RSS results}
\usage{
get_RSS(
  species = 21720,
  region = "goa",
  comp = "age",
  sex_cat = NULL,
  spec_case = NULL
)
}
\arguments{
\item{species}{AFSC species code (default = 21720, pacific cod)}

\item{region}{survey region. options are 'ai', 'ebs', 'ebs_slope', 'goa', and 'nebs' (default = 'goa')}

\item{comp}{type of composition for which RSS desired, options are 'age', 'length', and 'caal'}

\item{sex_cat}{sex category for which RSS desired, options are 0 (sexes combined pre-expansion), 1 (males), 2 (females), 12 (males-female comp that sums to 1), and 4 (sexes combined post-expansion) (default = NULL)}

\item{spec_case}{description string if getting RSS for special case, options are 'ai_subreg', 'bsre', 'dr', 'rebs', 'w_c_egoa', 'w140', 'wc_egoa' (default = NULL)}
}
\value{
a dataframe of bootstrap statistics for RSS
}
\description{
Function that retrieves composition replicates of Realized Sample Size (RSS) results for AFSC stock assessments
}
